requirements.txt
streamlit
pandas
numpy
scikit-surprise
scikit-learn
